MM-46246 : Move "License SKUs" type from MM-redux
