+++ components/admin_console/license_settings/trial_license_card/trial_license_card.tsx
@@ -1,23 +1,20 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 import React from 'react';
-
 import {FormattedMessage} from 'react-intl';
-
 import moment from 'moment';
 
 import {ClientLicense} from '@mattermost/types/config';
 
-import FormattedMarkdownMessage from 'components/formatted_markdown_message';
-
-import {daysToLicenseExpire} from 'utils/license_utils.jsx';
+import {daysToLicenseExpire} from 'utils/license_utils';
 import {getBrowserTimezone} from 'utils/timezone';
 
+import FormattedMarkdownMessage from 'components/formatted_markdown_message';
 import PurchaseLink from 'components/announcement_bar/purchase_link/purchase_link';
 import ContactUsButton from 'components/announcement_bar/contact_sales/contact_us';
+import AlertBanner from 'components/alert_banner';
 
 import './trial_license_card.scss';
-import AlertBanner from 'components/alert_banner';
 
 export interface Props {
     license: ClientLicense;
