+++ components/admin_console/permission_schemes_settings/permissions_tree/permissions_tree.test.tsx
@@ -4,7 +4,7 @@
 import React from 'react';
 import {shallow} from 'enzyme';
 
-import {LicenseSkus} from 'mattermost-redux/types/general';
+import {LicenseSkus} from 'utils/constants';
 
 import PermissionsTree from 'components/admin_console/permission_schemes_settings/permissions_tree/permissions_tree';
 import PermissionGroup from 'components/admin_console/permission_schemes_settings/permission_group';
