+++ components/admin_console/permission_schemes_settings/permissions_tree_playbooks.tsx
@@ -4,11 +4,12 @@
 import React from 'react';
 import {FormattedMessage} from 'react-intl';
 
-import Permissions from 'mattermost-redux/constants/permissions';
 import {ClientLicense} from '@mattermost/types/config';
-import {isEnterpriseLicense, isNonEnterpriseLicense} from 'mattermost-redux/types/general';
-
 import {Role} from '@mattermost/types/roles';
+
+import Permissions from 'mattermost-redux/constants/permissions';
+
+import {isEnterpriseLicense, isNonEnterpriseLicense} from 'utils/license_utils';
 
 import PermissionGroup from './permission_group';
