+++ utils/license_utils.test.ts
@@ -1,33 +1,34 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-import {isLicenseExpired, isLicenseExpiring, isLicensePastGracePeriod, isEnterpriseOrE20License} from 'utils/license_utils.jsx';
-import {LicenseSkus} from 'mattermost-redux/types/general';
+import {isLicenseExpired, isLicenseExpiring, isLicensePastGracePeriod, isEnterpriseOrE20License} from 'utils/license_utils';
+
+import {LicenseSkus} from 'utils/constants';
 
 describe('license_utils', () => {
     const millisPerDay = 24 * 60 * 60 * 1000;
     describe('isLicenseExpiring', () => {
         it('should return false if cloud expiring in 5 days', () => {
-            const license = {Id: '1234', IsLicensed: 'true', Cloud: 'true', ExpiresAt: Date.now() + (5 * millisPerDay)};
+            const license = {Id: '1234', IsLicensed: 'true', Cloud: 'true', ExpiresAt: `${Date.now() + (5 * millisPerDay)}`};
 
             expect(isLicenseExpiring(license)).toBeFalsy();
         });
 
         it('should return True if expiring in 5 days - non Cloud', () => {
-            const license = {Id: '1234', IsLicensed: 'true', Cloud: 'false', ExpiresAt: Date.now() + (5 * millisPerDay)};
+            const license = {Id: '1234', IsLicensed: 'true', Cloud: 'false', ExpiresAt: `${Date.now() + (5 * millisPerDay)}`};
 
             expect(isLicenseExpiring(license)).toBeTruthy();
         });
     });
     describe('isLicenseExpired', () => {
         it('should return false if cloud expired 1 day ago', () => {
-            const license = {Id: '1234', IsLicensed: 'true', Cloud: 'true', ExpiresAt: Date.now() - (Number(millisPerDay))};
+            const license = {Id: '1234', IsLicensed: 'true', Cloud: 'true', ExpiresAt: `${Date.now() - (Number(millisPerDay))}`};
 
             expect(isLicenseExpired(license)).toBeFalsy();
         });
 
         it('should return True if expired 1 day ago - non Cloud', () => {
-            const license = {Id: '1234', IsLicensed: 'true', Cloud: 'false', ExpiresAt: Date.now() - (Number(millisPerDay))};
+            const license = {Id: '1234', IsLicensed: 'true', Cloud: 'false', ExpiresAt: `${Date.now() - (Number(millisPerDay))}`};
 
             expect(isLicenseExpired(license)).toBeTruthy();
         });
@@ -35,13 +36,13 @@ describe('license_utils', () => {
 
     describe('isLicensePastGracePeriod', () => {
         it('should return False if cloud expired 11 days ago', () => {
-            const license = {Id: '1234', IsLicensed: 'true', Cloud: 'true', ExpiresAt: Date.now() - (11 * millisPerDay)};
+            const license = {Id: '1234', IsLicensed: 'true', Cloud: 'true', ExpiresAt: `${Date.now() - (11 * millisPerDay)}`};
 
             expect(isLicensePastGracePeriod(license)).toBeFalsy();
         });
 
         it('should return True if expired 1 day ago - non Cloud', () => {
-            const license = {Id: '1234', IsLicensed: 'true', Cloud: 'false', ExpiresAt: Date.now() - (11 * millisPerDay)};
+            const license = {Id: '1234', IsLicensed: 'true', Cloud: 'false', ExpiresAt: `${Date.now() - (11 * millisPerDay)}`};
 
             expect(isLicensePastGracePeriod(license)).toBeTruthy();
         });
